refactor(habitat): project D2 governance facts
