--- tools/habitat-harness/src/rules/registry.ts
@@ -1,4 +1,0 @@
-export { ruleGraphFacts } from "../plugin/rule-graph.js";
-export * from "./registry/load.js";
-export * from "./registry/projections.js";
-export * from "./registry/schema.js";
