+++ tools/habitat-harness/src/rules/registry/index.ts
@@ -1,0 +1,4 @@
+export { ruleGraphFacts } from "../../plugin/rule-graph.js";
+export * from "./load.js";
+export * from "./projections.js";
+export * from "./schema.js";
